# src/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cart API Using Inversify",
        "description": "Here I have created a cart api using IOC structure using Inversify Framework",
        "version": "1.0.0",
        "contact": {
            "name": "Rajput Chirag",
            "email": "csronly4@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Backend Server"
        },
        {
            "url": "http://localhost:4000",
            "description": "Frontend Server"
        }
    ],
    "paths": {
        "/api/signup": {
            "post": {
                "tags": ["User"],
                "summary": "Create a new user",
                "description": "Create a new user with the provided details",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": ["User"],
                "summary": "User login",
                "description": "Authenticate user with provided credentials",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "passworsd1d@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/protected": {
            "get": {
                "tags": ["Protected"],
                "summary": "Protected Path With Authentication",
                "description": "Protected Path Using JWT Authentication",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Protected Route"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/products": {
            "post": {
                "tags": ["Product"],
                "summary": "For Adding the products",
                "description": "Protected Path Using JWT Authentication for adding products",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Added Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401" : {
                        "description" : "Unauthorized"
                    }
                }
            }
        },


        "/api/products/getProductsByName": {
            "post": {
                "tags": ["Product"],
                "summary": "For Getting the products",
                "description": "Get The Products by names here names are unique",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pankha"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Fetched Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/products/getAllProducts": {
            "get": {
                "tags": ["Product"],
                "summary": "For Getting all the products",
                "description": "Get all the products by name",
                "responses": {
                    "200": {
                        "description": "Product Fetched Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/profile/AddProfile": {
            "post": {
                "tags": ["Profile"],
                "summary": "For Adding the profile",
                "description": "For Adding the profile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chirag"
                                    }
                                },
                                "required": [
                                   "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Added Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/profile/GetProfiles": {
            "get": {
                "tags": ["Profile"],
                "summary": "For Getting the profile",
                "description": "For Getting the profile",
                "responses": {
                    "200": {
                        "description": "Product Fetched Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/profile/GetProfilesByName": {
            "post": {
                "tags": ["Profile"],
                "summary": "For Getting the profile by name",
                "description": "For Getting the profile by name",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Chirag"
                                    }
                                },
                                "required": [
                                   "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Fetched Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    }
                }
            },
            "Profile" : {
                "type" : "object",
                "properties" : {
                    "name" : {
                        "type" : "string"
                    },
                    "user_id" : {
                        "type" : "string"
                 }
            }
        }
    },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}